.......PS. [ZBX-10892] We do not return response values in case of no response
The problem relied on the fact that we didn't check if a response time
> 0 was given. Due to the fact that a response time of 0 is impossible
even on a local machine, we check before setting the new value if the
response time is bigger than 0 and only in that case we set it.
If there was no response from the http server, we ignore the value.
In this way all the value that represent the response time can be correctly
averaged on the frontend, having less polluted data