.......PS. [ZBX-10892] Reverted previous commit and changed the step processing
A comment correctly pointed out that we have not only the response time value
on a web check, but also download speed and HTTP response code checks.
So i reverted the previous modification that was checking for the time only and
applied and tested the patch proposed on JIRA. With this we have the same
behavior for all the values and in case an HTTP server doesn't respond to a
request at all, we ignore all the related data. This is a more consistent behavior
in the case a web server doesn't give to zabbix an answer.

Comment and proposed fix by Andris Zeila. Thanks!