...G...... [ZBX-11568] First proposed fix for correct rounding
As there is no internal rules regarding interger to double roundings,
i assumed the classic case where we want to round to the nearest
integer.
So, for this we use the standard round() function, that take the result
of the division (inferred as double) and round it to the nearest value.