........S. [ZBX-11864] Removed IPv6 preprocessor directive
There's no purpose on checking for IPv6 support in this case as the
real checking is performed by LibCURL. We are bound anyway on
how the LibCURL library was compiled in the first place (with IPv6
support or not).
In the worst case, the check will fail and the proper error will be
reported in the logs.

Suggested by gleb. Thanks!