...G...... [ZBX-11902] fix the cpu count per-partition
The zabbix agent was previously counting the number of CPUs
on a global basis (machine-wide), instead of monitoring just
the LPAR (virtual machine) where it was running. Fix the count
by using the right APIs for that.
Out of the question remain the WPAR partitions for now, as
those share a pool of resources with the parent LPAR and are
more like containers or jails.